--- pkg/api/latest/dynakube/conversion.go
@@ -1,7 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package dynakube
-
-// Hub tags this version as the 'source' of the conversion for controller runtime.
-func (*DynaKube) Hub() {}
