+++ pkg/api/latest/dynakube/dynakube_props.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/Dynatrace/dynatrace-operator/pkg/api/conversion"
 	"github.com/Dynatrace/dynatrace-operator/pkg/api/exp"
+	"github.com/Dynatrace/dynatrace-operator/pkg/api/latest/dynakube/otlp"
 	"github.com/Dynatrace/dynatrace-operator/pkg/api/status"
 	"github.com/Dynatrace/dynatrace-operator/pkg/util/dtapiurl"
 	"github.com/Dynatrace/dynatrace-operator/pkg/util/kubernetes/fields/k8senv"
@@ -191,4 +192,11 @@ func (dk *DynaKube) GetResourceAttributes() map[string]string {
 
 func (dk *DynaKube) PublicRegistryOverride() string {
 	return dk.Spec.PublicRegistryOverride
+}
+
+// Hub tags this version as the 'source' of the conversion for controller runtime.
+func (*DynaKube) Hub() {}
+
+func (dk *DynaKube) OTLPExporterConfiguration() *otlp.ExporterConfiguration {
+	return otlp.NewExporterConfiguration(dk.Spec.OTLPExporterConfiguration, dk.GetResourceAttributes())
 }
