+++ pkg/api/latest/dynakube/metadataenrichment/spec.go
@@ -67,3 +67,9 @@ func IsSupportedType(ruleType RuleType) bool {
 
 	return false
 }
+
+// +kubebuilder:object:generate=true
+
+type Status struct {
+	Rules []Rule `json:"rules,omitempty"`
+}
