--- pkg/api/latest/dynakube/metadataenrichment/status.go
@@ -1,10 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package metadataenrichment
-
-// +kubebuilder:object:generate=true
-
-type Status struct {
-	Rules []Rule `json:"rules,omitempty"`
-}
