--- pkg/api/v1alpha2/edgeconnect/config.go
@@ -1,8 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package edgeconnect
-
-import "time"
-
-const DefaultMinRequestThreshold = 15 * time.Minute
