--- pkg/api/v1alpha2/edgeconnect/conversion.go
@@ -1,7 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package edgeconnect
-
-// Hub tags this version as the 'source' of the conversion for controller runtime.
-func (*EdgeConnect) Hub() {}
