+++ pkg/api/validation/edgeconnect/api_server_test.go
@@ -11,8 +11,9 @@ import (
 )
 
 const (
-	testName      = "test-name"
-	testNamespace = "test-namespace"
+	testName               = "test-name"
+	testNamespace          = "test-namespace"
+	testServiceAccountName = "test"
 )
 
 func TestApiServer(t *testing.T) {
