--- pkg/api/validation/edgeconnect/config.go
@@ -1,8 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package validation
-
-const (
-	testServiceAccountName = "test"
-)
