--- pkg/consts/agent_injection.go
@@ -1,11 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package consts
-
-const (
-	BootstrapperInitSecretName      = "dynatrace-bootstrapper-config"
-	BootstrapperInitCertsSecretName = "dynatrace-bootstrapper-certs"
-
-	AgentInitBinDirMount = "/mnt/bin"
-)
