--- pkg/consts/host_availability.go
@@ -1,6 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package consts
-
-const HostAvailabilityDetectionEnvVar = "DT_HOST_AVAILABILITY_DETECTION"
