--- pkg/consts/otlp.go
@@ -1,9 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package consts
-
-const (
-	OTLPExporterSecretName      = "dynatrace-otlp-exporter-config"
-	OTLPExporterCertsSecretName = "dynatrace-otlp-exporter-certs"
-)
