--- pkg/consts/paths.go
@@ -1,6 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package consts
-
-const DTComponentsSecretsRootDir = "/var/lib/dynatrace/secrets"
