--- pkg/controllers/dynakube/activegate/internal/statefulset/conditions.go
@@ -1,6 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package statefulset
-
-const ActiveGateStatefulSetConditionType string = "ActiveGateStatefulSet"
