--- pkg/controllers/dynakube/activegate/internal/statefulset/config.go
@@ -1,8 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package statefulset
-
-const (
-	InternalProxySecretVolumeName = "internal-proxy-secret-volume"
-)
