--- pkg/controllers/dynakube/dtpullsecret/conditions.go
@@ -1,6 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package dtpullsecret
-
-const PullSecretConditionType = "PullSecret"
