+++ pkg/controllers/dynakube/dtpullsecret/reconciler.go
@@ -20,7 +20,8 @@ import (
 )
 
 const (
-	PullSecretSuffix = "-pull-secret"
+	PullSecretSuffix        = "-pull-secret"
+	PullSecretConditionType = "PullSecret"
 )
 
 type Reconciler struct {
