--- pkg/controllers/dynakube/logmonitoring/configsecret/conditions.go
@@ -1,8 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package configsecret
-
-const (
-	LmcConditionType = "LogMonitoringConfig"
-)
