+++ pkg/controllers/dynakube/logmonitoring/configsecret/reconciler.go
@@ -26,6 +26,9 @@ import (
 )
 
 const (
+	LmcConditionType         = "LogMonitoringConfig"
+	DeploymentConfigFilename = "deployment.conf"
+
 	logMonitoringSecretSuffix = "-logmonitoring-config"
 
 	TokenHashAnnotationKey   = api.InternalFlagPrefix + "tenant-token-hash"
