--- pkg/controllers/dynakube/logmonitoring/daemonset/conditions.go
@@ -1,8 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package daemonset
-
-const (
-	ConditionType = "LogMonitoringDaemonSet"
-)
