+++ pkg/controllers/dynakube/logmonitoring/daemonset/reconciler.go
@@ -24,6 +24,8 @@ import (
 )
 
 const (
+	ConditionType = "LogMonitoringDaemonSet"
+
 	serviceAccountName = "dynatrace-logmonitoring"
 )
