--- pkg/injection/codemodule/installer/binary/config.go
@@ -1,8 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package binary
-
-const (
-	VersionLatest = "latest"
-)
