+++ pkg/injection/codemodule/installer/binary/installer.go
@@ -19,6 +19,8 @@ import (
 	"github.com/pkg/errors"
 )
 
+const VersionLatest = "latest"
+
 type Properties struct {
 	OS            string
 	Arch          string
