--- pkg/injection/codemodule/installer/symlink/config.go
@@ -1,10 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package symlink
-
-const (
-	// example match: 1.239.14.20220325-164521
-	versionRegexp = `^(\d+)\.(\d+)\.(\d+)\.(\d+)-(\d+)$`
-	binDir        = "/agent/bin"
-)
