--- pkg/injection/namespace/bootstrapperconfig/conditions.go
@@ -1,9 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package bootstrapperconfig
-
-const (
-	ConfigConditionType = "BootstrapperConfig"
-	CertsConditionType  = "BootstrapperCertsConfig"
-)
