--- pkg/injection/otlp/exporterconfig/conditions.go
@@ -1,9 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package exporterconfig
-
-const (
-	ConfigConditionType = "OTLPExporterConfigSecret"
-	CertsConditionType  = "OTLPExporterCertsConfig"
-)
