--- pkg/injection/otlp/exporterconfig/config.go
@@ -1,7 +1,0 @@
-// Copyright Dynatrace LLC
-// SPDX-License-Identifier: Apache-2.0
-
-package exporterconfig
-
-// ActiveGateCertDataName is the key used to store ActiveGate certificate data in the secret containing the ActiveGate Certificate for the OTLP exporter.
-const ActiveGateCertDataName = "activegate-tls.crt"
