+++ pkg/util/kubernetes/objects/k8sstatefulset/builder.go
@@ -6,12 +6,17 @@ package k8sstatefulset
 import (
 	"slices"
 
+	"github.com/Dynatrace/dynatrace-operator/pkg/api"
 	"github.com/Dynatrace/dynatrace-operator/pkg/util/hasher"
 	"github.com/Dynatrace/dynatrace-operator/pkg/util/kubernetes/objects/internal/builder"
 	maputils "github.com/Dynatrace/dynatrace-operator/pkg/util/map"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+const (
+	AnnotationPVCHash = api.InternalFlagPrefix + "pvc-hash"
 )
 
 type Option = builder.Option[*appsv1.StatefulSet]
